+++ TaskManager/Controllers/TagsController.cs
@@ -20,24 +20,31 @@ namespace TaskManager.Controllers
 
         [HttpGet]
         [Route("[action]")]
-        public async Task<IEnumerable<Tag>> GetAll()
+        public async Task<ActionResult<IEnumerable<Tag>>> GetAll()
         {
             return await _context.Tags.ToListAsync();
         }
 
         [HttpGet("{id}")]
         [Route("[action]/{id}")]
-        public async Task<Tag> GetDetails(int id)
+        public async Task<ActionResult<Tag>> GetDetails(int id)
         {
-            return await _context.Tags
+            Tag tag = await _context.Tags
                 .Where(tag => tag.Id == id)
-                    .Include(tag => tag.TagAssignments)
+                .Include(tag => tag.TagAssignments)
                 .SingleOrDefaultAsync();
+
+            if (tag == null)
+            {
+                return NotFound();
+            }
+
+            return tag;
         }
 
         [HttpGet("{id}")]
         [Route("[action]/{id}")]
-        public async Task<IEnumerable<Tag>> GetByTodo(int id)
+        public async Task<ActionResult<IEnumerable<Tag>>> GetByTodo(int id)
         {
             Todo todo = await _context.Todos
                 .Where(t => t.Id == id)
@@ -46,14 +53,39 @@ namespace TaskManager.Controllers
                         .ThenInclude(tag => tag.TagAssignments)
                 .SingleOrDefaultAsync();
 
+            if (todo == null)
+            {
+                return NotFound();
+            }
+
             return todo.TagAssignments.Select(ta => ta.Tag).ToList();
         }
 
         [HttpPut("{id}")]
         public async Task<ActionResult<Tag>> Put(int id, [FromBody]Tag tag)
         {
+            if (id != tag.Id)
+            {
+                return BadRequest();
+            }
+
             _context.Entry(tag).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (await _context.Tags.FindAsync(id) == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
 
             return Ok(tag);
         }
@@ -71,6 +103,12 @@ namespace TaskManager.Controllers
         public async Task<ActionResult<Tag>> Delete(int id)
         {
             Tag tag = await _context.Tags.FindAsync(id);
+
+            if (tag == null)
+            {
+                return NotFound();
+            }
+
             _context.Tags.Remove(tag);
             await _context.SaveChangesAsync();
